% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianquilts-package.R
\docType{package}
\name{bayesianquilts-package}
\alias{bayesianquilts}
\alias{bayesianquilts-package}
\title{bayesianquilts: Adaptive Importance Sampling and Quilted-Model Scaling}
\description{
Tools for interpretable Bayesian machine learning built around two ideas:
}
\details{
\itemize{
\item Adaptive Importance Sampling (AIS) for Leave-One-Out (LOO)
cross-validation, with small-step (LL/NLL/KL/NKL/Variance/PMM1-3) and
global (MM1-3/MixIS) transformations targeting finite-variance IS
weights when standard PSIS-LOO fails.
\item The additive interaction decomposition of multi-way regression
coefficients into orthogonal components, with the
generalization-preserving prior scales of Chang (2026) derived from
each component's effective sample size.
}

See \code{system.file("doc", "ais_loo.Rmd", package = "bayesianquilts")}
for a walkthrough of the AIS workflow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mederrata/bayesianquilts}
  \item Report bugs at \url{https://github.com/mederrata/bayesianquilts/issues}
}

}
\author{
\strong{Maintainer}: Joshua Chang \email{josh@mederrata.org}

Authors:
\itemize{
  \item Joshua Chang \email{josh@mederrata.org}
}

}
\keyword{internal}
