% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_utils.R
\name{BrmsLikelihood}
\alias{BrmsLikelihood}
\title{Stan-backed Likelihood Function for brms models}
\description{
Uses rstan::grad_log_prob() to compute gradients and numerical
Hessian diagonals via Stan Math autodiff, enabling AIS transforms
for any brms model family without hand-coded derivatives.
}
\section{Super class}{
\code{\link[bayesianquilts:LikelihoodFunction]{LikelihoodFunction}} -> \code{BrmsLikelihood}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-BrmsLikelihood-initialize}{\code{BrmsLikelihood$new()}}
    \item \href{#method-BrmsLikelihood-log_likelihood}{\code{BrmsLikelihood$log_likelihood()}}
    \item \href{#method-BrmsLikelihood-log_likelihood_gradient}{\code{BrmsLikelihood$log_likelihood_gradient()}}
    \item \href{#method-BrmsLikelihood-log_likelihood_hessian_diag}{\code{BrmsLikelihood$log_likelihood_hessian_diag()}}
    \item \href{#method-BrmsLikelihood-clone}{\code{BrmsLikelihood$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="LikelihoodFunction" data-id="total_log_likelihood"><a href='../../bayesianquilts/html/LikelihoodFunction.html#method-LikelihoodFunction-total_log_likelihood'><code>LikelihoodFunction$total_log_likelihood()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BrmsLikelihood-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-BrmsLikelihood-initialize}{}}}
\subsection{\code{BrmsLikelihood$new()}}{
  Initialize from a brmsfit
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BrmsLikelihood$new(fit)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{fit}}{A brmsfit object}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BrmsLikelihood-log_likelihood"></a>}}
\if{latex}{\out{\hypertarget{method-BrmsLikelihood-log_likelihood}{}}}
\subsection{\code{BrmsLikelihood$log_likelihood()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BrmsLikelihood$log_likelihood(data, params)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{Data object (list or environment)}
      \item{\code{params}}{List of parameters (S x K matrices or arrays)}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BrmsLikelihood-log_likelihood_gradient"></a>}}
\if{latex}{\out{\hypertarget{method-BrmsLikelihood-log_likelihood_gradient}{}}}
\subsection{\code{BrmsLikelihood$log_likelihood_gradient()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BrmsLikelihood$log_likelihood_gradient(data, params)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{Data object}
      \item{\code{params}}{List of parameters}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BrmsLikelihood-log_likelihood_hessian_diag"></a>}}
\if{latex}{\out{\hypertarget{method-BrmsLikelihood-log_likelihood_hessian_diag}{}}}
\subsection{\code{BrmsLikelihood$log_likelihood_hessian_diag()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BrmsLikelihood$log_likelihood_hessian_diag(data, params)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{Data object}
      \item{\code{params}}{List of parameters}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BrmsLikelihood-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BrmsLikelihood-clone}{}}}
\subsection{\code{BrmsLikelihood$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BrmsLikelihood$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
