% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{Dimension}
\alias{Dimension}
\title{Dimension of a multi-way interaction}
\description{
Wraps a name + either an integer cardinality or an explicit list of
category values.
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{name}}{Character. Variable name.}

    \item{\code{cardinality}}{Integer. Number of categories.}

    \item{\code{values}}{Vector of category values (defaults to 0:(cardinality-1)).}

    \item{\code{continuous}}{Logical (default FALSE).}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Dimension-initialize}{\code{Dimension$new()}}
    \item \href{#method-Dimension-print}{\code{Dimension$print()}}
    \item \href{#method-Dimension-clone}{\code{Dimension$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dimension-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Dimension-initialize}{}}}
\subsection{\code{Dimension$new()}}{
  Construct a Dimension.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Dimension$new(name, cardinality, continuous = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{Variable name.}
      \item{\code{cardinality}}{Integer count or a vector of explicit values.}
      \item{\code{continuous}}{Whether the dimension is continuous (default FALSE).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dimension-print"></a>}}
\if{latex}{\out{\hypertarget{method-Dimension-print}{}}}
\subsection{\code{Dimension$print()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Dimension$print(...)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dimension-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Dimension-clone}{}}}
\subsection{\code{Dimension$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Dimension$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
