% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{entropy}
\alias{entropy}
\title{Compute entropy of weights}
\usage{
entropy(weights)
}
\arguments{
\item{weights}{Numeric vector or matrix of weights}
}
\value{
Entropy value
}
\description{
Compute entropy of weights
}
