% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_utils.R
\name{extract_brms_data}
\alias{extract_brms_data}
\title{Extract data from a brmsfit object for AIS}
\usage{
extract_brms_data(fit)
}
\arguments{
\item{fit}{A brmsfit object}
}
\value{
List with X and y
}
\description{
Extract data from a brmsfit object for AIS
}
