% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_utils.R
\name{extract_brms_params}
\alias{extract_brms_params}
\title{Extract parameters from a brmsfit object for AIS}
\usage{
extract_brms_params(fit)
}
\arguments{
\item{fit}{A brmsfit object}
}
\value{
List of parameters (beta, intercept) suitable for AIS likelihoods
}
\description{
Extract parameters from a brmsfit object for AIS
}
