% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{KLDivergence}
\alias{KLDivergence}
\title{KL Divergence Transformation}
\description{
Q_i = -exp(log_pi - log_ell_i) * grad(log_ell_i)
}
\details{
Uses the analytical divergence
div(Q_KL) = c_i * (-tr(H_i) + (grad log_pi - grad log_ell_i) . grad log_ell_i)
where c_i = -exp(log_pi - log_ell_i) and grad log_pi ~ sum_j grad log_ell_j
(posterior-gradient approximation, ignoring prior). Matches the Python
zero-Jacobian fix.
}
\section{Super classes}{
\code{\link[bayesianquilts:Transformation]{Transformation}} -> \code{\link[bayesianquilts:SmallStepTransformation]{SmallStepTransformation}} -> \code{KLDivergence}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-KLDivergence-compute_Q}{\code{KLDivergence$compute_Q()}}
    \item \href{#method-KLDivergence-compute_divergence_Q}{\code{KLDivergence$compute_divergence_Q()}}
    \item \href{#method-KLDivergence-clone}{\code{KLDivergence$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="Transformation" data-id="compute_importance_weights"><a href='../../bayesianquilts/html/Transformation.html#method-Transformation-compute_importance_weights'><code>Transformation$compute_importance_weights()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="Transformation" data-id="compute_moments"><a href='../../bayesianquilts/html/Transformation.html#method-Transformation-compute_moments'><code>Transformation$compute_moments()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="Transformation" data-id="initialize"><a href='../../bayesianquilts/html/Transformation.html#method-Transformation-initialize'><code>Transformation$initialize()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="SmallStepTransformation" data-id="call"><a href='../../bayesianquilts/html/SmallStepTransformation.html#method-SmallStepTransformation-call'><code>SmallStepTransformation$call()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="SmallStepTransformation" data-id="normalize_vector_field"><a href='../../bayesianquilts/html/SmallStepTransformation.html#method-SmallStepTransformation-normalize_vector_field'><code>SmallStepTransformation$normalize_vector_field()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KLDivergence-compute_Q"></a>}}
\if{latex}{\out{\hypertarget{method-KLDivergence-compute_Q}{}}}
\subsection{\code{KLDivergence$compute_Q()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{KLDivergence$compute_Q(
  theta,
  data,
  params,
  current_log_ell,
  log_pi = NULL,
  ...
)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KLDivergence-compute_divergence_Q"></a>}}
\if{latex}{\out{\hypertarget{method-KLDivergence-compute_divergence_Q}{}}}
\subsection{\code{KLDivergence$compute_divergence_Q()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{KLDivergence$compute_divergence_Q(theta, data, params, current_log_ell, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{theta}}{Current parameter PyTree (list of S x N x K arrays)}
      \item{\code{data}}{Data object}
      \item{\code{params}}{Original parameters (S x K)}
      \item{\code{current_log_ell}}{Current log-likelihoods (S x N)}
      \item{\code{...}}{Extra args forwarded to compute_Q}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KLDivergence-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KLDivergence-clone}{}}}
\subsection{\code{KLDivergence$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{KLDivergence$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
