% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{LikelihoodDescent}
\alias{LikelihoodDescent}
\title{Likelihood Descent Transformation}
\description{
Q = -grad(log_ell). Moves samples in the direction that decreases
the likelihood of the left-out observation.
}
\details{
Uses the analytical divergence div(Q) = -tr(H) (Hessian diagonal sum),
matching the Python implementation. This avoids the zero-Jacobian bug
from numerical differencing when h*div(Q) is small.
}
\section{Super classes}{
\code{\link[bayesianquilts:Transformation]{Transformation}} -> \code{\link[bayesianquilts:SmallStepTransformation]{SmallStepTransformation}} -> \code{LikelihoodDescent}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-LikelihoodDescent-compute_Q}{\code{LikelihoodDescent$compute_Q()}}
    \item \href{#method-LikelihoodDescent-compute_divergence_Q}{\code{LikelihoodDescent$compute_divergence_Q()}}
    \item \href{#method-LikelihoodDescent-clone}{\code{LikelihoodDescent$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="Transformation" data-id="compute_importance_weights"><a href='../../bayesianquilts/html/Transformation.html#method-Transformation-compute_importance_weights'><code>Transformation$compute_importance_weights()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="Transformation" data-id="compute_moments"><a href='../../bayesianquilts/html/Transformation.html#method-Transformation-compute_moments'><code>Transformation$compute_moments()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="Transformation" data-id="initialize"><a href='../../bayesianquilts/html/Transformation.html#method-Transformation-initialize'><code>Transformation$initialize()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="SmallStepTransformation" data-id="call"><a href='../../bayesianquilts/html/SmallStepTransformation.html#method-SmallStepTransformation-call'><code>SmallStepTransformation$call()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="SmallStepTransformation" data-id="normalize_vector_field"><a href='../../bayesianquilts/html/SmallStepTransformation.html#method-SmallStepTransformation-normalize_vector_field'><code>SmallStepTransformation$normalize_vector_field()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LikelihoodDescent-compute_Q"></a>}}
\if{latex}{\out{\hypertarget{method-LikelihoodDescent-compute_Q}{}}}
\subsection{\code{LikelihoodDescent$compute_Q()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LikelihoodDescent$compute_Q(theta, data, params, current_log_ell, ...)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LikelihoodDescent-compute_divergence_Q"></a>}}
\if{latex}{\out{\hypertarget{method-LikelihoodDescent-compute_divergence_Q}{}}}
\subsection{\code{LikelihoodDescent$compute_divergence_Q()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LikelihoodDescent$compute_divergence_Q(
  theta,
  data,
  params,
  current_log_ell,
  ...
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{theta}}{Current parameter PyTree (list of S x N x K arrays)}
      \item{\code{data}}{Data object}
      \item{\code{params}}{Original parameters (S x K)}
      \item{\code{current_log_ell}}{Current log-likelihoods (S x N)}
      \item{\code{...}}{Extra args forwarded to compute_Q}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LikelihoodDescent-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LikelihoodDescent-clone}{}}}
\subsection{\code{LikelihoodDescent$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LikelihoodDescent$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
