% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ais_classes.R
\name{LikelihoodFunction}
\alias{LikelihoodFunction}
\title{Abstract Base Class for Likelihood Functions}
\description{
Defines the interface for likelihood functions used in AIS.
Users must implement these methods for their specific models.
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-LikelihoodFunction-log_likelihood}{\code{LikelihoodFunction$log_likelihood()}}
    \item \href{#method-LikelihoodFunction-log_likelihood_gradient}{\code{LikelihoodFunction$log_likelihood_gradient()}}
    \item \href{#method-LikelihoodFunction-log_likelihood_hessian_diag}{\code{LikelihoodFunction$log_likelihood_hessian_diag()}}
    \item \href{#method-LikelihoodFunction-total_log_likelihood}{\code{LikelihoodFunction$total_log_likelihood()}}
    \item \href{#method-LikelihoodFunction-clone}{\code{LikelihoodFunction$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LikelihoodFunction-log_likelihood"></a>}}
\if{latex}{\out{\hypertarget{method-LikelihoodFunction-log_likelihood}{}}}
\subsection{\code{LikelihoodFunction$log_likelihood()}}{
  Compute log-likelihood
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LikelihoodFunction$log_likelihood(data, params)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{Data object (list or environment)}
      \item{\code{params}}{List of parameters (S x K matrices or arrays)}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Matrix of log-likelihoods (S x N)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LikelihoodFunction-log_likelihood_gradient"></a>}}
\if{latex}{\out{\hypertarget{method-LikelihoodFunction-log_likelihood_gradient}{}}}
\subsection{\code{LikelihoodFunction$log_likelihood_gradient()}}{
  Compute gradient of log-likelihood
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LikelihoodFunction$log_likelihood_gradient(data, params)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{Data object}
      \item{\code{params}}{List of parameters}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    List of gradients (matching params structure, e.g. S x N x K)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LikelihoodFunction-log_likelihood_hessian_diag"></a>}}
\if{latex}{\out{\hypertarget{method-LikelihoodFunction-log_likelihood_hessian_diag}{}}}
\subsection{\code{LikelihoodFunction$log_likelihood_hessian_diag()}}{
  Compute diagonal of Hessian of log-likelihood
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LikelihoodFunction$log_likelihood_hessian_diag(data, params)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{Data object}
      \item{\code{params}}{List of parameters}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    List of hessian diagonals (matching params structure, e.g. S x N x K)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LikelihoodFunction-total_log_likelihood"></a>}}
\if{latex}{\out{\hypertarget{method-LikelihoodFunction-total_log_likelihood}{}}}
\subsection{\code{LikelihoodFunction$total_log_likelihood()}}{
  Compute total log-likelihood across ALL observations for per-obs params.

For per-observation transformed parameters with shape (S, N_params, K),
computes for each (s, i):
\code{result[s, i] = sum_j log l(d_j | params[s, i, :])}.

Default implementation loops over the N dimension.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LikelihoodFunction$total_log_likelihood(data, params)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{Data object}
      \item{\code{params}}{Per-observation transformed parameters}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Matrix of shape (S, N_params) with total log-likelihoods
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LikelihoodFunction-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LikelihoodFunction-clone}{}}}
\subsection{\code{LikelihoodFunction$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LikelihoodFunction$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
