% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{LinearRegressionLikelihood}
\alias{LinearRegressionLikelihood}
\title{Linear Regression Likelihood (Gaussian)}
\description{
Linear Regression Likelihood (Gaussian)
}
\section{Super class}{
\code{\link[bayesianquilts:LikelihoodFunction]{LikelihoodFunction}} -> \code{LinearRegressionLikelihood}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-LinearRegressionLikelihood-initialize}{\code{LinearRegressionLikelihood$new()}}
    \item \href{#method-LinearRegressionLikelihood-log_likelihood}{\code{LinearRegressionLikelihood$log_likelihood()}}
    \item \href{#method-LinearRegressionLikelihood-log_likelihood_gradient}{\code{LinearRegressionLikelihood$log_likelihood_gradient()}}
    \item \href{#method-LinearRegressionLikelihood-log_likelihood_hessian_diag}{\code{LinearRegressionLikelihood$log_likelihood_hessian_diag()}}
    \item \href{#method-LinearRegressionLikelihood-clone}{\code{LinearRegressionLikelihood$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="LikelihoodFunction" data-id="total_log_likelihood"><a href='../../bayesianquilts/html/LikelihoodFunction.html#method-LikelihoodFunction-total_log_likelihood'><code>LikelihoodFunction$total_log_likelihood()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LinearRegressionLikelihood-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-LinearRegressionLikelihood-initialize}{}}}
\subsection{\code{LinearRegressionLikelihood$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LinearRegressionLikelihood$new()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LinearRegressionLikelihood-log_likelihood"></a>}}
\if{latex}{\out{\hypertarget{method-LinearRegressionLikelihood-log_likelihood}{}}}
\subsection{\code{LinearRegressionLikelihood$log_likelihood()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LinearRegressionLikelihood$log_likelihood(data, params)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{Data object (list or environment)}
      \item{\code{params}}{List of parameters (S x K matrices or arrays)}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LinearRegressionLikelihood-log_likelihood_gradient"></a>}}
\if{latex}{\out{\hypertarget{method-LinearRegressionLikelihood-log_likelihood_gradient}{}}}
\subsection{\code{LinearRegressionLikelihood$log_likelihood_gradient()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LinearRegressionLikelihood$log_likelihood_gradient(data, params)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{Data object}
      \item{\code{params}}{List of parameters}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LinearRegressionLikelihood-log_likelihood_hessian_diag"></a>}}
\if{latex}{\out{\hypertarget{method-LinearRegressionLikelihood-log_likelihood_hessian_diag}{}}}
\subsection{\code{LinearRegressionLikelihood$log_likelihood_hessian_diag()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LinearRegressionLikelihood$log_likelihood_hessian_diag(data, params)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{Data object}
      \item{\code{params}}{List of parameters}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LinearRegressionLikelihood-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LinearRegressionLikelihood-clone}{}}}
\subsection{\code{LinearRegressionLikelihood$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LinearRegressionLikelihood$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
