% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{LogisticRegressionLikelihood}
\alias{LogisticRegressionLikelihood}
\title{Logistic Regression Likelihood}
\description{
Logistic Regression Likelihood
}
\section{Super class}{
\code{\link[bayesianquilts:LikelihoodFunction]{LikelihoodFunction}} -> \code{LogisticRegressionLikelihood}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-LogisticRegressionLikelihood-initialize}{\code{LogisticRegressionLikelihood$new()}}
    \item \href{#method-LogisticRegressionLikelihood-log_likelihood}{\code{LogisticRegressionLikelihood$log_likelihood()}}
    \item \href{#method-LogisticRegressionLikelihood-log_likelihood_gradient}{\code{LogisticRegressionLikelihood$log_likelihood_gradient()}}
    \item \href{#method-LogisticRegressionLikelihood-log_likelihood_hessian_diag}{\code{LogisticRegressionLikelihood$log_likelihood_hessian_diag()}}
    \item \href{#method-LogisticRegressionLikelihood-clone}{\code{LogisticRegressionLikelihood$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="LikelihoodFunction" data-id="total_log_likelihood"><a href='../../bayesianquilts/html/LikelihoodFunction.html#method-LikelihoodFunction-total_log_likelihood'><code>LikelihoodFunction$total_log_likelihood()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegressionLikelihood-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegressionLikelihood-initialize}{}}}
\subsection{\code{LogisticRegressionLikelihood$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LogisticRegressionLikelihood$new()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegressionLikelihood-log_likelihood"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegressionLikelihood-log_likelihood}{}}}
\subsection{\code{LogisticRegressionLikelihood$log_likelihood()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LogisticRegressionLikelihood$log_likelihood(data, params)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{Data object (list or environment)}
      \item{\code{params}}{List of parameters (S x K matrices or arrays)}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegressionLikelihood-log_likelihood_gradient"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegressionLikelihood-log_likelihood_gradient}{}}}
\subsection{\code{LogisticRegressionLikelihood$log_likelihood_gradient()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LogisticRegressionLikelihood$log_likelihood_gradient(data, params)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{Data object}
      \item{\code{params}}{List of parameters}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegressionLikelihood-log_likelihood_hessian_diag"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegressionLikelihood-log_likelihood_hessian_diag}{}}}
\subsection{\code{LogisticRegressionLikelihood$log_likelihood_hessian_diag()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LogisticRegressionLikelihood$log_likelihood_hessian_diag(data, params)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{Data object}
      \item{\code{params}}{List of parameters}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegressionLikelihood-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegressionLikelihood-clone}{}}}
\subsection{\code{LogisticRegressionLikelihood$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LogisticRegressionLikelihood$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
