% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logSumExp}
\alias{logSumExp}
\title{Compute log-sum-exp safely}
\usage{
logSumExp(x)
}
\arguments{
\item{x}{Numeric vector or matrix}
}
\value{
Numeric value or vector
}
\description{
Compute log-sum-exp safely
}
