% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{MixIS}
\alias{MixIS}
\title{MixIS - Mixture Importance Sampling for LOO-CV (Silva & Zanella 2024)}
\description{
Resamples posterior draws using the mixture proposal
q_mix(theta) ~ pi(theta|D) * sum_i 1/ell(theta|d_i)
and reweights with finite-variance IS weights
nu_i(theta*) = (1/ell(theta*|d_i)) / sum_j (1/ell(theta*|d_j)).
Provides a robust LOO baseline that does not require posterior smoothing.
}
\section{Super class}{
\code{\link[bayesianquilts:Transformation]{Transformation}} -> \code{MixIS}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-MixIS-initialize}{\code{MixIS$new()}}
    \item \href{#method-MixIS-call}{\code{MixIS$call()}}
    \item \href{#method-MixIS-clone}{\code{MixIS$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="Transformation" data-id="compute_importance_weights"><a href='../../bayesianquilts/html/Transformation.html#method-Transformation-compute_importance_weights'><code>Transformation$compute_importance_weights()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="Transformation" data-id="compute_moments"><a href='../../bayesianquilts/html/Transformation.html#method-Transformation-compute_moments'><code>Transformation$compute_moments()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixIS-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-MixIS-initialize}{}}}
\subsection{\code{MixIS$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MixIS$new(likelihood_fn, n_mix_samples = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{likelihood_fn}}{LikelihoodFunction object}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixIS-call"></a>}}
\if{latex}{\out{\hypertarget{method-MixIS-call}{}}}
\subsection{\code{MixIS$call()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MixIS$call(
  max_iter,
  params,
  theta,
  data,
  log_ell,
  log_ell_original = NULL,
  log_pi = NULL,
  variational = FALSE,
  surrogate_log_prob_fn = NULL,
  seed = 42,
  ...
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{max_iter}}{Integer, maximum iterations}
      \item{\code{params}}{List of parameters}
      \item{\code{theta}}{Top-level parameters (usually same as params)}
      \item{\code{data}}{Data object}
      \item{\code{log_ell}}{Matrix of log-likelihoods}
      \item{\code{...}}{Additional arguments}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixIS-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MixIS-clone}{}}}
\subsection{\code{MixIS$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MixIS$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
