% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{MM1}
\alias{MM1}
\title{MM1 (Moment Matching 1) - Global shift transformation}
\description{
MM1 (Moment Matching 1) - Global shift transformation
}
\section{Super class}{
\code{\link[bayesianquilts:Transformation]{Transformation}} -> \code{MM1}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-MM1-call}{\code{MM1$call()}}
    \item \href{#method-MM1-clone}{\code{MM1$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="Transformation" data-id="compute_importance_weights"><a href='../../bayesianquilts/html/Transformation.html#method-Transformation-compute_importance_weights'><code>Transformation$compute_importance_weights()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="Transformation" data-id="compute_moments"><a href='../../bayesianquilts/html/Transformation.html#method-Transformation-compute_moments'><code>Transformation$compute_moments()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="Transformation" data-id="initialize"><a href='../../bayesianquilts/html/Transformation.html#method-Transformation-initialize'><code>Transformation$initialize()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MM1-call"></a>}}
\if{latex}{\out{\hypertarget{method-MM1-call}{}}}
\subsection{\code{MM1$call()}}{
  Apply MM1 transformation
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MM1$call(
  max_iter,
  params,
  theta,
  data,
  log_ell,
  log_ell_original = NULL,
  log_pi = NULL,
  variational = FALSE,
  surrogate_log_prob_fn = NULL,
  ...
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{max_iter}}{Integer, maximum iterations}
      \item{\code{params}}{List of parameters}
      \item{\code{theta}}{Top-level parameters (usually same as params)}
      \item{\code{data}}{Data object}
      \item{\code{log_ell}}{Matrix of log-likelihoods}
      \item{\code{...}}{Additional arguments}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MM1-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MM1-clone}{}}}
\subsection{\code{MM1$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MM1$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
