% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{NaturalKLDivergence}
\alias{NaturalKLDivergence}
\title{Natural-gradient KL Divergence Transformation}
\description{
Preconditions the KL gradient by the posterior covariance matrix,
aligning the step direction with the influence-function optimal shift.

Q = Sigma @ Q_kl, where Sigma is the posterior covariance and
Q_kl is the standard KL divergence vector field.
}
\section{Super classes}{
\code{\link[bayesianquilts:Transformation]{Transformation}} -> \code{\link[bayesianquilts:SmallStepTransformation]{SmallStepTransformation}} -> \code{NaturalKLDivergence}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-NaturalKLDivergence-initialize}{\code{NaturalKLDivergence$new()}}
    \item \href{#method-NaturalKLDivergence-compute_Q}{\code{NaturalKLDivergence$compute_Q()}}
    \item \href{#method-NaturalKLDivergence-compute_divergence_Q}{\code{NaturalKLDivergence$compute_divergence_Q()}}
    \item \href{#method-NaturalKLDivergence-clone}{\code{NaturalKLDivergence$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="Transformation" data-id="compute_importance_weights"><a href='../../bayesianquilts/html/Transformation.html#method-Transformation-compute_importance_weights'><code>Transformation$compute_importance_weights()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="Transformation" data-id="compute_moments"><a href='../../bayesianquilts/html/Transformation.html#method-Transformation-compute_moments'><code>Transformation$compute_moments()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="SmallStepTransformation" data-id="call"><a href='../../bayesianquilts/html/SmallStepTransformation.html#method-SmallStepTransformation-call'><code>SmallStepTransformation$call()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="SmallStepTransformation" data-id="normalize_vector_field"><a href='../../bayesianquilts/html/SmallStepTransformation.html#method-SmallStepTransformation-normalize_vector_field'><code>SmallStepTransformation$normalize_vector_field()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaturalKLDivergence-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-NaturalKLDivergence-initialize}{}}}
\subsection{\code{NaturalKLDivergence$new()}}{
  Initialize
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{NaturalKLDivergence$new(likelihood_fn, posterior_cov)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{likelihood_fn}}{LikelihoodFunction object}
      \item{\code{posterior_cov}}{Posterior covariance matrix (K_total x K_total)}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaturalKLDivergence-compute_Q"></a>}}
\if{latex}{\out{\hypertarget{method-NaturalKLDivergence-compute_Q}{}}}
\subsection{\code{NaturalKLDivergence$compute_Q()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{NaturalKLDivergence$compute_Q(
  theta,
  data,
  params,
  current_log_ell,
  log_pi = NULL,
  ...
)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaturalKLDivergence-compute_divergence_Q"></a>}}
\if{latex}{\out{\hypertarget{method-NaturalKLDivergence-compute_divergence_Q}{}}}
\subsection{\code{NaturalKLDivergence$compute_divergence_Q()}}{
  Analytical divergence of Q_NKL = Sigma \%\emph{\% Q_KL.
div(Q_NKL) = c_i * (-tr(Sigma \%}\% H_i) + (sum_j g_j - g_i)^T Sigma g_i)
with diagonal-Hessian approximation for tr(Sigma \%*\% H).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{NaturalKLDivergence$compute_divergence_Q(
  theta,
  data,
  params,
  current_log_ell,
  ...
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{theta}}{Current parameter PyTree (list of S x N x K arrays)}
      \item{\code{data}}{Data object}
      \item{\code{params}}{Original parameters (S x K)}
      \item{\code{current_log_ell}}{Current log-likelihoods (S x N)}
      \item{\code{...}}{Extra args forwarded to compute_Q}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaturalKLDivergence-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NaturalKLDivergence-clone}{}}}
\subsection{\code{NaturalKLDivergence$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{NaturalKLDivergence$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
