% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{NaturalLikelihoodDescent}
\alias{NaturalLikelihoodDescent}
\title{Natural-gradient Likelihood Descent Transformation}
\description{
\code{Q_NLL = Sigma \%*\% (-grad log ell_i)}, where \code{Sigma} is the
posterior covariance. Approximates the influence function
\code{Jinv \%*\% (grad log ell_i)} without the \code{pi(theta | D)} prefactor of
the KL flow. Cheaper than KL (no posterior evaluation) and more directly
aligned with the optimal LOO shift.

Analytical divergence: div(Q_NLL) = -tr(Sigma \%*\% H) with diagonal-Hessian
approximation: -sum_k Sigma_kk * H_kk.
}
\section{Super classes}{
\code{\link[bayesianquilts:Transformation]{Transformation}} -> \code{\link[bayesianquilts:SmallStepTransformation]{SmallStepTransformation}} -> \code{NaturalLikelihoodDescent}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-NaturalLikelihoodDescent-initialize}{\code{NaturalLikelihoodDescent$new()}}
    \item \href{#method-NaturalLikelihoodDescent-compute_Q}{\code{NaturalLikelihoodDescent$compute_Q()}}
    \item \href{#method-NaturalLikelihoodDescent-compute_divergence_Q}{\code{NaturalLikelihoodDescent$compute_divergence_Q()}}
    \item \href{#method-NaturalLikelihoodDescent-clone}{\code{NaturalLikelihoodDescent$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="Transformation" data-id="compute_importance_weights"><a href='../../bayesianquilts/html/Transformation.html#method-Transformation-compute_importance_weights'><code>Transformation$compute_importance_weights()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="Transformation" data-id="compute_moments"><a href='../../bayesianquilts/html/Transformation.html#method-Transformation-compute_moments'><code>Transformation$compute_moments()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="SmallStepTransformation" data-id="call"><a href='../../bayesianquilts/html/SmallStepTransformation.html#method-SmallStepTransformation-call'><code>SmallStepTransformation$call()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="SmallStepTransformation" data-id="normalize_vector_field"><a href='../../bayesianquilts/html/SmallStepTransformation.html#method-SmallStepTransformation-normalize_vector_field'><code>SmallStepTransformation$normalize_vector_field()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaturalLikelihoodDescent-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-NaturalLikelihoodDescent-initialize}{}}}
\subsection{\code{NaturalLikelihoodDescent$new()}}{
  Initialize
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{NaturalLikelihoodDescent$new(likelihood_fn, posterior_cov)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{likelihood_fn}}{LikelihoodFunction object}
      \item{\code{posterior_cov}}{Posterior covariance matrix (K_total x K_total)}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaturalLikelihoodDescent-compute_Q"></a>}}
\if{latex}{\out{\hypertarget{method-NaturalLikelihoodDescent-compute_Q}{}}}
\subsection{\code{NaturalLikelihoodDescent$compute_Q()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{NaturalLikelihoodDescent$compute_Q(theta, data, params, current_log_ell, ...)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaturalLikelihoodDescent-compute_divergence_Q"></a>}}
\if{latex}{\out{\hypertarget{method-NaturalLikelihoodDescent-compute_divergence_Q}{}}}
\subsection{\code{NaturalLikelihoodDescent$compute_divergence_Q()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{NaturalLikelihoodDescent$compute_divergence_Q(
  theta,
  data,
  params,
  current_log_ell,
  ...
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{theta}}{Current parameter PyTree (list of S x N x K arrays)}
      \item{\code{data}}{Data object}
      \item{\code{params}}{Original parameters (S x K)}
      \item{\code{current_log_ell}}{Current log-likelihoods (S x N)}
      \item{\code{...}}{Extra args forwarded to compute_Q}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaturalLikelihoodDescent-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NaturalLikelihoodDescent-clone}{}}}
\subsection{\code{NaturalLikelihoodDescent$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{NaturalLikelihoodDescent$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
