% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{PMM2}
\alias{PMM2}
\title{PMM2 (Partial Moment Matching 2) - Scale + Shift SmallStep transformation}
\description{
Q = (ratio - 1) * val + (mean_w - ratio * mean)
where ratio = sqrt(var_w / var)
}
\section{Super classes}{
\code{\link[bayesianquilts:Transformation]{Transformation}} -> \code{\link[bayesianquilts:SmallStepTransformation]{SmallStepTransformation}} -> \code{PMM2}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-PMM2-compute_Q}{\code{PMM2$compute_Q()}}
    \item \href{#method-PMM2-compute_divergence_Q}{\code{PMM2$compute_divergence_Q()}}
    \item \href{#method-PMM2-clone}{\code{PMM2$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="Transformation" data-id="compute_importance_weights"><a href='../../bayesianquilts/html/Transformation.html#method-Transformation-compute_importance_weights'><code>Transformation$compute_importance_weights()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="Transformation" data-id="compute_moments"><a href='../../bayesianquilts/html/Transformation.html#method-Transformation-compute_moments'><code>Transformation$compute_moments()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="Transformation" data-id="initialize"><a href='../../bayesianquilts/html/Transformation.html#method-Transformation-initialize'><code>Transformation$initialize()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="SmallStepTransformation" data-id="call"><a href='../../bayesianquilts/html/SmallStepTransformation.html#method-SmallStepTransformation-call'><code>SmallStepTransformation$call()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="SmallStepTransformation" data-id="normalize_vector_field"><a href='../../bayesianquilts/html/SmallStepTransformation.html#method-SmallStepTransformation-normalize_vector_field'><code>SmallStepTransformation$normalize_vector_field()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMM2-compute_Q"></a>}}
\if{latex}{\out{\hypertarget{method-PMM2-compute_Q}{}}}
\subsection{\code{PMM2$compute_Q()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PMM2$compute_Q(
  theta,
  data,
  params,
  current_log_ell,
  log_ell_original = NULL,
  ...
)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMM2-compute_divergence_Q"></a>}}
\if{latex}{\out{\hypertarget{method-PMM2-compute_divergence_Q}{}}}
\subsection{\code{PMM2$compute_divergence_Q()}}{
  Compute divergence of PMM2's Q.
Q_k = (ratio_k - 1) * theta_k + const, so dQ_k/dtheta_k = ratio_k - 1.
div(Q) = sum_k (ratio_k - 1).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PMM2$compute_divergence_Q(theta, data, params, current_log_ell, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{theta}}{Current parameter PyTree (list of S x N x K arrays)}
      \item{\code{data}}{Data object}
      \item{\code{params}}{Original parameters (S x K)}
      \item{\code{current_log_ell}}{Current log-likelihoods (S x N)}
      \item{\code{...}}{Extra args forwarded to compute_Q}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMM2-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PMM2-clone}{}}}
\subsection{\code{PMM2$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PMM2$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
