% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{PMM3}
\alias{PMM3}
\title{PMM3 (Partial Moment Matching 3) - full-rank affine with step size}
\description{
Generalization of MM3 with a tunable step size \code{h}. The vector field is
\code{Q(theta) = (L_w \%*\% Linv - I) \%*\% (theta - mu) + (mu_w - mu)},
so \code{T(theta) = theta + h * Q(theta)} interpolates between identity
(\code{h=0}) and full MM3 (\code{h=1}). Divergence is exact:
\code{div(Q) = trace(L_w \%*\% Linv - I)}, constant in \code{theta}.
}
\section{Super classes}{
\code{\link[bayesianquilts:Transformation]{Transformation}} -> \code{\link[bayesianquilts:SmallStepTransformation]{SmallStepTransformation}} -> \code{PMM3}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-PMM3-compute_Q}{\code{PMM3$compute_Q()}}
    \item \href{#method-PMM3-compute_divergence_Q}{\code{PMM3$compute_divergence_Q()}}
    \item \href{#method-PMM3-clone}{\code{PMM3$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="Transformation" data-id="compute_importance_weights"><a href='../../bayesianquilts/html/Transformation.html#method-Transformation-compute_importance_weights'><code>Transformation$compute_importance_weights()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="Transformation" data-id="compute_moments"><a href='../../bayesianquilts/html/Transformation.html#method-Transformation-compute_moments'><code>Transformation$compute_moments()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="Transformation" data-id="initialize"><a href='../../bayesianquilts/html/Transformation.html#method-Transformation-initialize'><code>Transformation$initialize()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="SmallStepTransformation" data-id="call"><a href='../../bayesianquilts/html/SmallStepTransformation.html#method-SmallStepTransformation-call'><code>SmallStepTransformation$call()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="SmallStepTransformation" data-id="normalize_vector_field"><a href='../../bayesianquilts/html/SmallStepTransformation.html#method-SmallStepTransformation-normalize_vector_field'><code>SmallStepTransformation$normalize_vector_field()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMM3-compute_Q"></a>}}
\if{latex}{\out{\hypertarget{method-PMM3-compute_Q}{}}}
\subsection{\code{PMM3$compute_Q()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PMM3$compute_Q(
  theta,
  data,
  params,
  current_log_ell,
  log_ell_original = NULL,
  ...
)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMM3-compute_divergence_Q"></a>}}
\if{latex}{\out{\hypertarget{method-PMM3-compute_divergence_Q}{}}}
\subsection{\code{PMM3$compute_divergence_Q()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PMM3$compute_divergence_Q(theta, data, params, current_log_ell, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{theta}}{Current parameter PyTree (list of S x N x K arrays)}
      \item{\code{data}}{Data object}
      \item{\code{params}}{Original parameters (S x K)}
      \item{\code{current_log_ell}}{Current log-likelihoods (S x N)}
      \item{\code{...}}{Extra args forwarded to compute_Q}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PMM3-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PMM3-clone}{}}}
\subsection{\code{PMM3$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PMM3$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
