% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psis.R
\name{psislw}
\alias{psislw}
\title{Pareto Smoothed Importance Sampling (Wrapper around loo)}
\usage{
psislw(log_ratios)
}
\arguments{
\item{log_ratios}{Numeric matrix of log importance ratios (Draws x Observations).}
}
\value{
A list with elements:
\item{log_weights}{Matrix of smoothed log weights}
\item{weights}{Matrix of smoothed normalized weights}
\item{khat}{Vector of Pareto k diagnostics}
}
\description{
Pareto Smoothed Importance Sampling (Wrapper around loo)
}
