% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tree_map}
\alias{tree_map}
\title{Map a function over a nested list (simple PyTree equivalent)}
\usage{
tree_map(f, x, ...)
}
\arguments{
\item{f}{Function to apply}

\item{x}{List or atomic object}

\item{...}{Additional arguments to f}
}
\value{
Transformed list or object
}
\description{
Map a function over a nested list (simple PyTree equivalent)
}
