% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unflatten_pytree_leaves}
\alias{unflatten_pytree_leaves}
\title{Inverse of flatten_pytree_leaves.}
\usage{
unflatten_pytree_leaves(flat_arr, layout)
}
\arguments{
\item{flat_arr}{Array of shape (S, N, K_total).}

\item{layout}{As returned by flatten_pytree_leaves.}
}
\value{
Named list of arrays matching the original PyTree.
}
\description{
Inverse of flatten_pytree_leaves.
}
\keyword{internal}
