% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{Variance}
\alias{Variance}
\title{Variance-based Transformation}
\description{
Q = pi * (f/ell)^2 * grad(log(f/ell))
= pi * exp(2 log f - 2 log ell) * (grad log f - grad log ell)

For the default target f = ell, this collapses to grad log f - grad log ell
which is zero, so callers typically supply a custom f_fn (e.g. an
expectation target).

Analytical divergence (with default f = ell, so delta_g = 0):
div(Q_Var) ~ w * (-tr(H) - 2 * ||grad log ell||^2)
}
\section{Super classes}{
\code{\link[bayesianquilts:Transformation]{Transformation}} -> \code{\link[bayesianquilts:SmallStepTransformation]{SmallStepTransformation}} -> \code{Variance}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Variance-initialize}{\code{Variance$new()}}
    \item \href{#method-Variance-compute_Q}{\code{Variance$compute_Q()}}
    \item \href{#method-Variance-compute_divergence_Q}{\code{Variance$compute_divergence_Q()}}
    \item \href{#method-Variance-clone}{\code{Variance$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="Transformation" data-id="compute_importance_weights"><a href='../../bayesianquilts/html/Transformation.html#method-Transformation-compute_importance_weights'><code>Transformation$compute_importance_weights()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="Transformation" data-id="compute_moments"><a href='../../bayesianquilts/html/Transformation.html#method-Transformation-compute_moments'><code>Transformation$compute_moments()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="SmallStepTransformation" data-id="call"><a href='../../bayesianquilts/html/SmallStepTransformation.html#method-SmallStepTransformation-call'><code>SmallStepTransformation$call()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bayesianquilts" data-topic="SmallStepTransformation" data-id="normalize_vector_field"><a href='../../bayesianquilts/html/SmallStepTransformation.html#method-SmallStepTransformation-normalize_vector_field'><code>SmallStepTransformation$normalize_vector_field()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Variance-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Variance-initialize}{}}}
\subsection{\code{Variance$new()}}{
  Initialize
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Variance$new(likelihood_fn, f_fn = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{likelihood_fn}}{LikelihoodFunction}
      \item{\code{f_fn}}{Optional function(data, params) -> (S, N) target.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Variance-compute_Q"></a>}}
\if{latex}{\out{\hypertarget{method-Variance-compute_Q}{}}}
\subsection{\code{Variance$compute_Q()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Variance$compute_Q(theta, data, params, current_log_ell, log_pi = NULL, ...)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Variance-compute_divergence_Q"></a>}}
\if{latex}{\out{\hypertarget{method-Variance-compute_divergence_Q}{}}}
\subsection{\code{Variance$compute_divergence_Q()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Variance$compute_divergence_Q(theta, data, params, current_log_ell, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{theta}}{Current parameter PyTree (list of S x N x K arrays)}
      \item{\code{data}}{Data object}
      \item{\code{params}}{Original parameters (S x K)}
      \item{\code{current_log_ell}}{Current log-likelihoods (S x N)}
      \item{\code{...}}{Extra args forwarded to compute_Q}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Variance-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Variance-clone}{}}}
\subsection{\code{Variance$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Variance$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
